extract common mma helpers to _mma module
